+++ packages/network/lib/hooks.dart
@@ -2,7 +2,5 @@ export 'src/exception.dart';
 export 'src/request.dart';
 export 'src/response.dart';
 export 'src/settings.dart';
-export 'src/network.dart';
-export 'src/middleware.dart';
-// TODO(Serge) remove in 1.0.0
 export 'src/hooks.dart';
+export 'src/middleware.dart';